Fix TPU Spawn gather (#6896)
